+++ payzen/classes/PayzenTools.php
@@ -12,6 +12,9 @@ if (! defined('_PS_VERSION_')) {
     exit;
 }
 
+use Lyranetwork\Payzen\Sdk\Form\Api as PayzenApi;
+use Lyranetwork\Payzen\Sdk\Refund\WsException as PayzenWsException;
+
 class PayzenTools
 {
     private static $GATEWAY_CODE = 'PayZen';
@@ -30,7 +33,7 @@ class PayzenTools
 
     private static $CMS_IDENTIFIER = 'PrestaShop_1.5-1.7';
     private static $SUPPORT_EMAIL = 'support@payzen.eu';
-    private static $PLUGIN_VERSION = '1.15.6';
+    private static $PLUGIN_VERSION = '1.15.7';
     private static $GATEWAY_VERSION = 'V2';
 
     const ORDER_ID_REGEX = '#^[a-zA-Z0-9]{1,9}$#';